food_prediction/requirements.txt
flask
flask_sqlalchemy
flask_login
pillow
torch
torchvision
tqdm
numpy
pandas
scikit-learn
gunicorn
kaggle
python-multipart